Fix login for invited users
When someone who is not already an exercism user gets
invited to join a team, we create a placeholder user for them containing
just their GitHub username.

The first time they log in, we associate their GitHub information (id,
avatar, email if available) with this placeholder user, and correct the
case of their username.

We were failing to store the github id, which would make it impossible
to log in, since we would be storing a nil github_id in the session.

Closes #2355.
Closes #2351.
Closes #2331.