Delete deprecated endpoints
1. demo

A very old version of the CLI called /assignments/demo which then
redirected to /demo, which then forwarded the call to x-api /demo,
which later was replaced with x-api /problems/demo... which is now
called directly from the CLI.

2. deprecated fetch endpoints

The CLI used to call endpoints in the exercism.io api in order
to deliver problems. Then we moved these into the x-api, but x-api
still needed to know what the user should have access to, so
they became a callback. Then we changed the endpoint paths to make
more sense, so the deprecated ones redirected to the new ones.

3. current

I can't even remember how this was used. Presumably it was a
way to get the problems that are active for a user right now,
which would let us return those problems during a 'fetch'.

4. peek

We no longer have the peek command in the CLI. If someone has an old
enough exercism binary that they have the peek command it's ok for it to
fail. We'll help them upgrade.